Moves ProxyResponder button to a new class
