+++ Xamarin.PropertyEditing.Mac/Controls/Custom/FocusableBooleanButton.cs
@@ -3,28 +3,6 @@ using AppKit;
 
 namespace Xamarin.PropertyEditing.Mac
 {
-	internal class ProxyResponderButton : NSButton
-	{
-		public RowProxyResponder ProxyResponder { get; set; }
-
-		public override void KeyDown (NSEvent theEvent)
-		{
-			switch (theEvent.KeyCode) {
-			case (int)NSKey.Tab:
-				if (ProxyResponder != null) {
-					if (theEvent.ModifierFlags.HasFlag(NSEventModifierMask.ShiftKeyMask)) {
-						ProxyResponder.PreviousResponder ();
-					} else {
-						ProxyResponder.NextResponder ();
-					}
-					return;
-				}
-				break;
-			}
-			base.KeyDown (theEvent);
-		}
-	}
-
 	internal class FocusableBooleanButton : ProxyResponderButton
 	{
 		public override bool CanBecomeKeyView { get { return Enabled; } }
